refactor(product page): Delete cross sell section
