--- frontend/models/product/sections/cross-sell-section.ts
@@ -1,9 +1,0 @@
-import { z } from 'zod';
-
-export type CrossSellSection = z.infer<typeof CrossSellSectionSchema>;
-
-export const CrossSellSectionSchema = z.object({
-   type: z.literal('CrossSell'),
-   id: z.string(),
-   title: z.string().nullable(),
-});
